Convert Safe stack simulation test to bun:test
